Move cleaned_data to data_io #283
